fix: refactor widgets layout
